+++ Music Browse/Assets.xcassets/mn.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Black-violin-iphone-wallpaper-ilikewallpaper_com.jpeg",
+      "filename" : "music-notes-red-background-illustration-white-music-notes-flowing-curved-stave-red-background-grunge-watercolor-157330460.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
